fix copyright
split CAN tests. Tests of cansockets can not be executed on a build server

load_layer('can') doesn't work on the build servers, somehow

try to remove CANSocket unit tests from regression tests

split CAN and CANSocket into different files

revert a previous test

remove buggy unit test

change unit test configuration to exclude cansocket.uts

fix OSX UT
